# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NSE/BSE AI Trading Signals API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get the current watchlist
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add a symbol to the watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWatchlistBody"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /watchlist/{id}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove a symbol from the watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    WatchlistItem:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        exchange:
          type: string
          enum: [NSE, BSE]
        instrumentType:
          type: string
          enum: [STOCK, INDEX, OPTIONS, FUTURES]
        addedAt:
          type: string
      required: [id, symbol, name, exchange, instrumentType, addedAt]
    AddWatchlistBody:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        exchange:
          type: string
          enum: [NSE, BSE]
          default: NSE
        instrumentType:
          type: string
          enum: [STOCK, INDEX, OPTIONS, FUTURES]
          default: STOCK
      required: [symbol, name]
...